requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.0
SQLAlchemy>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
transformers>=4.57.1
torch>=2.3.0
safetensors>=0.4.0
accelerate>=0.33.0
huggingface_hub>=0.24.0
python-dotenv>=1.0.0
protobuf>=4.25.0